+++ server/src/main/java/org/opensearch/snapshots/RestoreService.java
@@ -382,7 +382,7 @@ public class RestoreService implements ClusterStateApplier {
 
                                 IndexId snapshotIndexId = repositoryData.resolveIndexId(index);
 
-                                final Settings overrideSettingsInternal = getOverrideSettingsInternal();
+                                final Settings overrideSettingsInternal = getOverrideSettingsInternal(metadata.index(index));
                                 final String[] ignoreSettingsInternal = getIgnoreSettingsInternal();
 
                                 IndexMetadata snapshotIndexMetadata = updateIndexSettings(
@@ -464,12 +464,12 @@ public class RestoreService implements ClusterStateApplier {
                                     IndexMetadata.Builder indexMdBuilder = IndexMetadata.builder(snapshotIndexMetadata)
                                         .state(IndexMetadata.State.OPEN)
                                         .index(renamedIndexName);
-                                    indexMdBuilder.settings(
-                                        Settings.builder()
-                                            .put(snapshotIndexMetadata.getSettings())
-                                            .put(IndexMetadata.SETTING_INDEX_UUID, UUIDs.randomBase64UUID())
-                                    );
-                                    createIndexService.addRemoteStoreCustomMetadata(indexMdBuilder, false);
+                                    Settings idxSettings = Settings.builder()
+                                        .put(snapshotIndexMetadata.getSettings())
+                                        .put(IndexMetadata.SETTING_INDEX_UUID, UUIDs.randomBase64UUID())
+                                        .build();
+                                    indexMdBuilder.settings(idxSettings);
+                                    createIndexService.addRemoteStoreCustomMetadata(indexMdBuilder, idxSettings, false);
                                     shardLimitValidator.validateShardLimit(
                                         renamedIndexName,
                                         snapshotIndexMetadata.getSettings(),
@@ -688,7 +688,7 @@ public class RestoreService implements ClusterStateApplier {
                         return indexSettingsToBeIgnored;
                     }
 
-                    private Settings getOverrideSettingsInternal() {
+                    private Settings getOverrideSettingsInternal(IndexMetadata indexMetadata) {
                         final Settings.Builder settingsBuilder = Settings.builder();
 
                         // We will use whatever replication strategy provided by user or from snapshot metadata unless
@@ -712,7 +712,8 @@ public class RestoreService implements ClusterStateApplier {
                             clusterService.state(),
                             clusterSettings,
                             clusterService.getSettings(),
-                            String.join(",", request.indices())
+                            String.join(",", request.indices()),
+                            indexMetadata
                         );
                         return settingsBuilder.build();
                     }